Refactor configuration
* Added .env.example,
  - exposes the configuration vars that need sharing between CI and test/deployer

* Updated .gitignore
  - removed environments.js

* In contract-addresses.js,
  - updated the config blocks to hold only public endpoints, its all we need to track

* In environments.js,
  - refactored to pull in items from .env

* Removed environments_template.js

* In hardhat.config.js,
  - required dotenv

* In manage-roles.js and role-assignments.js,
  - updated paths

* In package.json,
  - replaced rinkeby with ropsten
  - added dotenv

* In setup.md,
  - updated configure environment section
