+++ .env.example
@@ -1,0 +1,19 @@
+# Gas limit override if default gas fails
+DEPLOYER_GAS_LIMIT=2045000
+
+# Etherscan API for verifying contracts on etherscan
+DEPLOYER_ETHERSCAN_API_KEY=000000000000000000000000000000000000000
+
+# Coinmarketcap API for reporting up to date gas price estimates for deployment and function calls
+GAS_REPORTER_COINMARKETCAP_API_KEY=000000000000000000000000000000000000000
+
+# Hardhat specific
+DEPLOYER_HARDHAT_MNEMONIC="reduce chair insane vault universe fitness flame motor wood toy vacuum special"
+
+# Ropsten specific 
+DEPLOYER_ROPSTEN_TXNODE=http://ropsten.node.example:666
+DEPLOYER_ROPSTEN_KEY=0xb644aa4a6ada1f1a996def5cee46b93922006d78754095525cefeed36f5ab1c0
+
+# Mainnet specific 
+DEPLOYER_MAINNET_TXNODE=http://mainnet.node.example:666
+DEPLOYER_MAINNET_KEY=0xb644aa4a6ada1f1a996def5cee46b93922006d78754095525cefeed36f5ab1c0
