+++ .gitignore
@@ -8,7 +8,6 @@ node_modules
 
 # Secrets
 .env
-.env.*
 
 # OS
 .DS_Store
@@ -24,6 +23,3 @@ node_modules
 
 # Test Coverage
 coverage*
-
-# Environment Config
-environments.js
