+++ environments.js
@@ -1,46 +1,47 @@
 /**
- *  Build and test environment configuration template.
+ *  Build and test environment configuration.
  *
- *  - Copy to environments.js and edit to suit local needs.
- *  - environments.js is in .gitignore and will not be committed
+ *  - Translates environment vars into JSON objects
+ *  - Environment vars are defined in .env and
  */
+
+require('dotenv').config()
+
 module.exports = {
 
     // For helping public deployments succeed if default gas doesn't work
-    "gasLimit": 20450000,
+    "gasLimit": parseInt(process.env.DEPLOYER_GAS_LIMIT),
 
     // Needed for verifying contract code on Etherscan
     "etherscan": {
-        "apiKey": ""
+        "apiKey": process.env.ETHERSCAN_API_KEY
     },
 
     // Needed for Gas Reporter
     "coinmarketcap": {
-        "apiKey": ""
+        "apiKey": process.env.GAS_REPORTER_COINMARKETCAP_API_KEY
     },
-
 
     // Hardhat testnet
     //  - throwaway HDWallet mnemonic for running unit tests, which require more than one address
     "hardhat": {
-        "txNode": "",
-        "mnemonic": "reduce chair insane vault universe fitness flame motor wood toy vacuum special"
+        "mnemonic": process.env.DEPLOYER_HARDHAT_MNEMONIC
     },
 
     // Ropsten testnet
     //  - placeholder private key is first address of test HDWallet used in hardhat network config
     //  - Replace key with multisig pk for deployment
     "ropsten": {
-        "txNode": "",
-        "keys": ["0xb644aa4a6ada1f1a996def5cee46b93922006d78754095525cefeed36f5ab1c0"]
+        "txNode": process.env.DEPLOYER_ROPSTEN_TXNODE,
+        "keys": [process.env.DEPLOYER_ROPSTEN_KEY]
     },
 
     // Ethereum Mainnet
     //  - placeholder private key is first address of test HDWallet used in hardhat network config
     //  - Replace key with multisig pk for deployment
     "mainnet": {
-        "txNode": "",
-        "keys": ["0xb644aa4a6ada1f1a996def5cee46b93922006d78754095525cefeed36f5ab1c0"]
+        "txNode": process.env.DEPLOYER_MAINNET_TXNODE,
+        "keys": [process.env.DEPLOYER_MAINNET_KEY]
     }
 
 };
