+++ hardhat.config.js
@@ -1,3 +1,6 @@
+const dotEnvConfig = require('dotenv');
+dotEnvConfig.config();
+
 const environments = require('./environments');
 require("@nomiclabs/hardhat-etherscan");
 require("@nomiclabs/hardhat-waffle");
@@ -5,6 +8,7 @@ require("@nomiclabs/hardhat-web3");
 require('hardhat-contract-sizer');
 require("hardhat-gas-reporter");
 require("solidity-coverage");
+
 
 module.exports = {
   defaultNetwork: "hardhat",
